Move flag_matching tests to test_flag_matching.rs
